requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
ftfy==6.3.1
geoip2==5.1.0
idna==3.10
maxminddb==2.8.2
multidict==6.6.3
mysqlclient==2.2.7
propcache==0.3.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
yarl==1.20.1
